Move minio module to drenv
Promote ramenctl.minio helper module to drenv.minio so we can use it in
the velero addon. This is module is not really generic and depend on the
specific way we deploy the minio addon, but we don't have a better place
for this now.
